Fix omnibus_install?
omnibus_install? was returning whether 'current ruby exe path', up three
levels existed as a valid path.  Because it's based on where ruby was
running, this will return true in most cases (unless you're running from
an executable in your home dir or somilar, '/home/blah/ruby').
It doesn't really answer the question 'is the running command running
in the context of an omnibus install'.

This check was modified so that we also look for the omnibus-generated
version-manifest.json file in the expected install path.   The
combination of tese two things answers the question above.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>