Prototype chef test command to provide backwards compat for delivery local mode
Long term goal:

Provide a one stop shop test command that just does the right thing out of the box. Abstract away the underlying tools so that a user doesn't need to know individual commands and can run just a single command to test their infrastructure. This might sound a bit familiar since this was the goal of Delivery Local Mode.

How did Delivery Local Mode Work?

Delivery Local Mode provided a simple config in .delivery/project.toml that defined commands to run for each of a predefined set of phases. It also allowed you define a URL where a config was located so all your cookbooks could use a single remote config that was easy to update. You ran your tests with delivery local PHASE or delivery local all to run all phases.

Short term goal of this PR:

Stop shipping the delivery command in Chef Workstation, while providing compatibility for users utilizing the Delivery Local Mode workflow. This PR lacks the actual shellout, debug logging, or any error handling, but it shows the concept of how this would work from the user's perspective. A user could run one of the following commands:
- chef test all: run all phases in the project.toml file
- chef test PHASE: run a single phase in the project.toml file
- chef test: run all phases (all is implied) in the project.toml file

If a remote config is defined in the local project.toml file that will be fetched and parsed as well.

Next step:

- Add the actual shellout
- Add error handling
- Add debug logging
- Add tests
- Add a shim delivery command to chef-cli. This command will act as a pass-through if a delivery command is found in the path, but otherwise it'll provide full compatibility for users running the legacy `delivery local` command

From here we can implement a v1 test solution (delivery local mode being v0) with its own config. This PR code would be the fallback legacy mode when a delivery config was found. The code would live on for some period of time while it warned users to upgrade to the new testing method.

Signed-off-by: Tim Smith <tsmith@chef.io>