Remove the `chef verify` hidden internal command
This command was used for CI/internal validation of the
chef-workstation package and the tools it ships.  It now lives
in the Chef Workstation repository as an omnibus pipeline validation
script `chef-workstation/omnibus/verification.rb`; which is invoked
by buildkite CI through `chef-workstation/omnibus/omnibus-test.ps1|sh`;
as configured in chef-workstation/.expeditor/release.omnibus.yml.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>