Run knife rehash each time we install a gem
knife rehash greatly speeds up the performance of knife by caching the location of all knife plugins in the .chef directory. With this change we'll rehash the list any time someone installs a gem into chef, as they may be installing a knife plugin.

Note: This is going to require changes to the subcommand loader that aren't yet in master of chef/chef so we'll be depending on a specific new chef version in the gemspec if we merge this.

Signed-off-by: Tim Smith <tsmith@chef.io>