Update the content of the generated chefignore
Add a few files used in all the sous-chefs cookbooks since a lot of people follow their examples and will have those same CI setups. This also removes the changelog since `knife supermarket share` is using the ignore file and that results in cookbooks without a changelog on the supermarket.

Signed-off-by: Tim Smith <tsmith@chef.io>