Add habitat to version output
In addition:
* add fallback versioning-  when a component
does not exist as a gem we try to resolve it out of the
package manifest.  This allows the caller to be unconcerned with
which manifest contains a particular component.
* remove workstation-like output from `chef-cli --version` in absence
of Workstation, so that when installed as a gem it doesn't try to
report on components that won't be installed.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>