Simplify and speed up how we detect if we're in an omnibus install
This gets called at the start of every command to the chef CLI. It involves a bunch of path discover, expansion, and file checks that aren't necessary. We can just take the directory that we're in and see if it starts with our omnibus directory structure. This is significantly faster and uses less memory according to stackprof profiling.

Signed-off-by: Tim Smith <tsmith@chef.io>