Fix Chef Workstation issue #1292
Issue #1292: chef install requires .rb file present #1292:

Issue located in the Chef CLI tool. The install.rb script only checked
for the presence of a policy file with an .rb extension whether or not
the user entered a policy lockfile. Changed code to check for .rb and
.lock.json extensions in the policy file path, and then verify the
existence of a policy file with that extension.

Updated error message in storage_config_spec.rb to correctly address
both .rb and .lock.json extensions.

Created test for new error language, and created test to verify error
when a policy file name is given, but the file does not exist.

DCO:
By making a contribution to this project, I certify that:

(a) The contribution was created in whole or in part by me and I
    have the right to submit it under the open source license
    indicated in the file; or

(b) The contribution is based upon previous work that, to the
    best of my knowledge, is covered under an appropriate open
    source license and I have the right under that license to
    submit that work with modifications, whether created in whole
    or in part by me, under the same open source license (unless
    I am permitted to submit under a different license), as
    Indicated in the file; or

(c) The contribution was provided directly to me by some other
    person who certified (a), (b) or (c) and I have not modified
    it.

(d) I understand and agree that this project and the contribution
    are public and that a record of the contribution (including
    all personal information I submit with it, including my
    sign-off) is maintained indefinitely and may be redistributed
    consistent with this project or the open source license(s)
    involved.

Signed-off-by: Rick Morrow <rclark.morrow@gmail.com>